Refactor security group rule processing
