+++ src/steps/resource-manager/network/converters/securityGroups.test.ts
@@ -13,7 +13,6 @@ import { SECURITY_GROUP_RULE_RELATIONSHIP_TYPE } from '../constants';
 import {
   createSecurityGroupRuleRelationships,
   createSecurityGroupRuleRelationshipsFromRule,
-  getPortsFromRange,
   processSecurityGroupRule,
 } from './securityGroups';
 
@@ -297,31 +296,5 @@ describe('build mapped relationships from security group rules', () => {
         _integrationInstanceId,
       ).length,
     ).toEqual(4);
-  });
-});
-
-describe('get port range from string', () => {
-  test('range', () => {
-    const portRange = '8080-8081';
-    expect(getPortsFromRange(portRange)).toEqual({
-      fromPort: 8080,
-      toPort: 8081,
-    });
-  });
-
-  test('single port', () => {
-    const portRange = '22';
-    expect(getPortsFromRange(portRange)).toEqual({
-      fromPort: 22,
-      toPort: 22,
-    });
-  });
-
-  test('* => 0-65535', () => {
-    const portRange = '*';
-    expect(getPortsFromRange(portRange)).toEqual({
-      fromPort: 0,
-      toPort: 65535,
-    });
   });
 });
