+++ src/steps/resource-manager/network/converters/types.ts
@@ -1,0 +1,9 @@
+export type SecurityRulePortRange = {
+  /**
+   * The original port range value.
+   */
+  portRange: string;
+
+  fromPort?: number;
+  toPort?: number;
+};
